// api-gateway/src/services/cars/dtos/car.dto.ts
export class CarModelDTO {
  name: string;
  luggagePayload: number;
  guestQuantity: number;
  partnerId: string;
  photoUrl: string;
  standardPricePerKm: number;
  country: string;
  city: string;
  classId: string;
  details: string;
}

export class CarDTO {
  licencePlate: string;
  color: string;
}